Fix all clippy lints for clean -D warnings
Resolve clippy lints across the workspace:
- collapsible_if: use let-chains instead of nested ifs
- borrow_deref_ref: remove redundant &* on AtomicU32::from_ptr
- derivable_impls: derive Default instead of manual impl
- manual_div_ceil: use .div_ceil()
- io_other_error: use std::io::Error::other()
- should_implement_trait: impl FromStr instead of inherent from_str
- len_without_is_empty: add is_empty() to ContentStore traits
- too_many_arguments: extract context structs or allow where appropriate
- unnecessary_cast, redundant_closure, large_enum_variant, dead_code

+++ crates/ndn-face-l2/src/ndrv.rs
@@ -224,7 +224,7 @@ impl NdrvSocket {
         }
 
         // Register EtherType 0x8624
-        let ethertype_be = (NDN_ETHERTYPE as u16).to_be_bytes();
+        let ethertype_be = NDN_ETHERTYPE.to_be_bytes();
         let demux = NdrvDemuxDesc {
             desc_type: NDRV_DEMUXTYPE_ETHERTYPE,
             desc_len: 2,
@@ -326,7 +326,7 @@ impl NdrvSocket {
         let mut frame = Vec::with_capacity(ETHER_HEADER_LEN + payload.len());
         frame.extend_from_slice(dst_mac.as_bytes());
         frame.extend_from_slice(self.local_mac.as_bytes());
-        let et = (NDN_ETHERTYPE as u16).to_be_bytes();
+        let et = NDN_ETHERTYPE.to_be_bytes();
         frame.extend_from_slice(&et);
         frame.extend_from_slice(payload);
